refactor(releases): minor final tidy to global perspective menu; creating release type section component
